fix: a bug fix
The path was containing a optional value that's mean creating the path in
this form Optional('\(path)'). to fix the bug unwrap the optinal value.         The second bug was: when we get the information from the server we cast them to a Bool type , it's always faild because the type was as [String:AnyObject] , to fix the bug we casted  to [String: AnyObject]